Add files via upload
Our project is movies and TV shows web-based adaptive
real-time recommendation system. It is a website hosted locally which asks the user to select a
minimum of 10 movies and rate them on a scale of 1-5. The system is based on collaborative
filtering and uses singular value decomposition (SVD). SVD worked best for the application of
various algorithms we tested.

Developed collaboratively with Raphael, Nischal, and Swapnil. 

Code Link: https://github.com/raphaelcfernandes/movieRecommender